Move SnapshotTransferService
